Rename Cli project to PS2MapTool.Cli Rename core library to PS2MapTool
