+++ PS2MapTool.Cli/Properties/launchSettings.json
