--- PS2MapTool.Core/Areas/NoDeployType.cs
@@ -1,8 +1,0 @@
-﻿namespace PS2MapTool.Core.Areas
-{
-    public enum NoDeployType
-    {
-        Sunderer = 200,
-        Ant = 2336
-    }
-}
