--- PS2MapTool.Core/Lod.cs
@@ -1,10 +1,0 @@
-﻿namespace PS2MapTool.Core
-{
-    public enum Lod
-    {
-        Lod0 = 0,
-        Lod1 = 1,
-        Lod2 = 2,
-        Lod3 = 3
-    }
-}
