--- PS2MapTool.Core/World.cs
@@ -1,16 +1,0 @@
-﻿namespace PS2MapTool.Core
-{
-    public enum World
-    {
-        Amerish,
-        Desolation,
-        Esamir,
-        Hossin,
-        Indar,
-        Koltyr,
-        Nexus,
-        Sanctuary,
-        Tutorial,
-        VR
-    }
-}
