+++ PS2MapTool.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31229.75
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PS2MapTool", "PS2MapTool\PS2MapTool.csproj", "{FFA42262-6C68-407A-915C-958CB10EE6E1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PS2MapTool.Cli", "PS2MapTool.Cli\PS2MapTool.Cli.csproj", "{FFA42262-6C68-407A-915C-958CB10EE6E1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PS2MapTool.Core", "PS2MapTool.Core\PS2MapTool.Core.csproj", "{BA9B0BA9-FD1B-41BB-B094-D7BCDB8D7689}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PS2MapTool", "PS2MapTool\PS2MapTool.csproj", "{BA9B0BA9-FD1B-41BB-B094-D7BCDB8D7689}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
