+++ PS2MapTool/Areas/NoDeployType.cs
@@ -1,0 +1,8 @@
+﻿namespace PS2MapTool.Areas
+{
+    public enum NoDeployType
+    {
+        Sunderer = 200,
+        Ant = 2336
+    }
+}
