+++ PS2MapTool/Lod.cs
@@ -1,0 +1,10 @@
+﻿namespace PS2MapTool
+{
+    public enum Lod
+    {
+        Lod0 = 0,
+        Lod1 = 1,
+        Lod2 = 2,
+        Lod3 = 3
+    }
+}
