+++ PS2MapTool/PS2MapTool.csproj
@@ -1,27 +1,14 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>net5.0</TargetFramework>
-    <Platforms>x64</Platforms>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="CliFx" Version="2.0.4" />
     <PackageReference Include="Pfim" Version="0.10.0" />
     <PackageReference Include="SixLabors.ImageSharp" Version="1.0.3" />
-    <PackageReference Include="Spectre.Console" Version="0.39.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\PS2MapTool.Core\PS2MapTool.Core.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="optipng.exe">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
+    <PackageReference Include="SixLabors.ImageSharp.Drawing" Version="1.0.0-beta13" />
   </ItemGroup>
 
 </Project>
