--- PS2MapTool/Program.cs
@@ -1,17 +1,0 @@
-﻿using CliFx;
-using System.Threading.Tasks;
-
-namespace PS2MapTool
-{
-    public static class Program
-    {
-        public static async Task<int> Main() =>
-            await new CliApplicationBuilder()
-                .SetDescription("Tools to extract and stitch together PlanetSide 2 world map LODs.")
-                .SetTitle("PS2 Map Tools")
-                .AddCommandsFromThisAssembly()
-                .Build()
-                .RunAsync()
-                .ConfigureAwait(false);
-    }
-}
