+++ PS2MapTool/World.cs
@@ -1,0 +1,16 @@
+﻿namespace PS2MapTool
+{
+    public enum World
+    {
+        Amerish,
+        Desolation,
+        Esamir,
+        Hossin,
+        Indar,
+        Koltyr,
+        Nexus,
+        Sanctuary,
+        Tutorial,
+        VR
+    }
+}
